requirements.txt
defusedxml==0.4.1
Django==1.10.1
django-braces==1.9.0
django-embed-video==1.1.0
django-mssql==1.7
django-pyodbc==0.3.0
django-pyodbc-azure==1.10.0.1
djangorestframework==3.4.5
mutagenx==1.22
oauthlib==1.1.2
Pillow==3.3.1
PyJWT==1.4.2
pyodbc==3.0.10
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
sql-server.pyodbc==1.0
